move export method to the builder
