+++ src/js/builder.js
@@ -1,3 +1,6 @@
+import JSZip from 'jszip';
+import saveAs from 'save-as'
+import { getImageBlob } from './util';
 import Section from './section';
 import installComponents from './components';
 
@@ -13,6 +16,7 @@ let _builder = null;
 export default class Builder {
   constructor (options) {
     options = Object.assign({}, BUILDER_OPTIONS, options);
+    this.title = options.docTitle;
     this.isEditing = true;
     this.sections = options.sections || [];
   }
@@ -68,5 +72,96 @@ export default class Builder {
 
     installComponents(Vue);
     _Vue = Vue;
+  }
+
+  _exportZip () {
+    const printPreview = window.open('about:blank', 'print_preview');
+    const printDocument = printPreview.document;
+    printDocument.write(
+      `<!DOCTYPE html>
+        <html>
+          ${document.documentElement.innerHTML}
+        </html>`
+    );
+    printDocument.open();
+    const editable = Array.from(printDocument.querySelectorAll('.is-editable'));
+    const uploder = Array.from(printDocument.querySelectorAll('.is-uploader'));
+    const stylers = Array.from(printDocument.querySelectorAll('.styler'));
+    const images = Array.from(printDocument.querySelectorAll('img'));
+    const artboadrd = printDocument.querySelector('#artboard');
+    const head = printDocument.querySelector('head');
+    const imagePromises = [];
+    const zip = new JSZip();
+    const output = zip.folder('project');
+    const imgFolder = output.folder('assets/img');
+
+    images.forEach((image) => {
+      const imageLoader = getImageBlob(image.src);
+      imagePromises.push(imageLoader);
+      imageLoader.then((img) => {
+        imgFolder.file(img.name, img.blob, { base64: true });
+        image.setAttribute('src', `assets/img/${img.name}`);
+      })
+    });
+
+    Promise.all(imagePromises).then(() => {
+      editable.forEach((el) => {
+        el.contentEditable = 'false';
+        el.classList.remove('is-editable');
+      });
+      uploder.forEach((el) => {
+        const input = el.querySelector(':scope > input');
+        input.remove();
+        el.classList.remove('is-uploader');
+      });
+      stylers.forEach((styler) => {
+        styler.remove();
+      });
+
+      output.file('index.html',
+        `<html>
+          <head>
+            ${head.innerHTML}
+          </head>
+          <body>
+            ${artboadrd.innerHTML}
+          </body>
+        </html>`);
+
+      zip.generateAsync({ type: 'blob' })
+        .then((blob) => {
+          saveAs(blob, 'project.zip');
+          printPreview.close();
+        });
+    });
+  }
+
+  toJSON () {
+    return {
+      title: this.docTitle,
+      sections: this.sections
+    }
+  }
+
+  preview () {
+    const printPreview = window.open('about:blank', 'print_preview');
+    const printDocument = printPreview.document;
+    printDocument.write(
+      `<!DOCTYPE html>
+        <html>
+          ${document.documentElement.innerHTML}
+        </html>`
+    );
+    printDocument.open();
+  }
+
+  export (method = 'zip') {
+    if (method === 'zip') {
+      return this._exportZip();
+    }
+
+    if (method === 'json') {
+      return this.toJSON();
+    }
   }
 };
