+++ src/pug/layout.pug
@@ -6,6 +6,6 @@ head
   meta(http-equiv="X-UA-Compatible", content="ie=edge")
   title Document
 body
-  include mixin/icon.pug
+  include mixin/icon
   #app
     block content
